Add hooks to menu building to allow filtering & tweaks
I've succesfully used these the filter hook for creating
a separate submenu for templates and another for sys-.* VMs.

I haven't played with the menuitem tweaking opportunities
a replacement qubes.make_menuitem would offer but as it was
easy to allow as a by-product of this change I see no harm
done allowing also that and it makes things more consistent.